+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 4.1.0 - 2026-04-12
+
+### Added
+
+- `Innmind\Framework\Application::mapRoutes()`
+
+### Changed
+
+- Reduced routes memory footprint
+
 ## 4.0.2 - 2026-04-12
 
 ### Fixed